Makes getEmployeeFromOption work with real types instead of optional
Usually it's uncommon to have the need to pass an Optional type
as parameter to a function/method. The Optional type is used to
comunicate that a specific value is nullable. But when you're
passing parameters to a function, you're in control of that,
so there's not really a need to comunicate to somebody else.